+++ CHANGELOG.md
@@ -10,6 +10,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.12.0] (2025-05-06)
+
+### Added
+- Added localization support.
+- Added support for rendering chapter markers & label.
+
+### Fixed
+
+- Fixed an issue where the play button would stay in the paused state after starting the stream.
+
 ## [0.11.0] (2025-04-03)
 
 ### Added